+++ Notificare/NotificarePushLib/Models/NotificareAttachment.h
@@ -11,8 +11,8 @@
 @interface NotificareAttachment : NSObject
 
 
-@property (strong, nonatomic) NSString * attachmentURI;
-@property (strong, nonatomic) NSString * attachmentMimeType;
+@property (strong, nonatomic, nonnull) NSString * attachmentURI;
+@property (strong, nonatomic, nonnull) NSString * attachmentMimeType;
 
 
 @end