+++ Notificare/NotificarePushLib/Models/NotificareBeacon.h
@@ -12,21 +12,21 @@
 @interface NotificareBeacon : NSObject
 
 
-@property (strong, nonatomic) NSString * name;
-@property (strong, nonatomic) NSString * region;
-@property (strong, nonatomic) NSUUID * beaconUUID;
-@property (strong, nonatomic) NSString * major;
-@property (strong, nonatomic) NSString * minor;
-@property (strong, nonatomic) NSString * proximity;
-@property (strong, nonatomic) NSDictionary * proximityNotifications;
-@property (strong, nonatomic) NSString * application;
-@property (strong, nonatomic) NSDictionary * notification;
-@property (strong, nonatomic) NSString * latitude;
-@property (strong, nonatomic) NSString * longitude;
-@property (strong, nonatomic) NSString * batteryLevel;
-@property (strong, nonatomic) NSString * purpose;
-@property (strong, nonatomic) CLBeacon * beacon;
-@property (strong, nonatomic) NSString * timezone;
+@property (strong, nonatomic, nonnull) NSString * name;
+@property (strong, nonatomic, nonnull) NSString * region;
+@property (strong, nonatomic, nonnull) NSUUID * beaconUUID;
+@property (strong, nonatomic, nonnull) NSString * major;
+@property (strong, nonatomic, nonnull) NSString * minor;
+@property (strong, nonatomic, nonnull) NSString * proximity;
+@property (strong, nonatomic, nullable) NSDictionary * proximityNotifications;
+@property (strong, nonatomic, nonnull) NSString * application;
+@property (strong, nonatomic, nonnull) NSDictionary * notification;
+@property (strong, nonatomic, nonnull) NSString * latitude;
+@property (strong, nonatomic, nonnull) NSString * longitude;
+@property (strong, nonatomic, nullable) NSString * batteryLevel;
+@property (strong, nonatomic, nonnull) NSString * purpose;
+@property (strong, nonatomic, nullable) CLBeacon * beacon;
+@property (strong, nonatomic, nonnull) NSString * timezone;
 @property (assign, nonatomic) BOOL triggers;
 
 @end