+++ Notificare/NotificarePushLib/Models/NotificareProduct.h
@@ -12,18 +12,18 @@
 @interface NotificareProduct : NSObject
 
 
-@property (strong, nonatomic) NSString * productName;
-@property (strong, nonatomic) NSString * productDescription;
-@property (strong, nonatomic) NSString * type;
-@property (strong, nonatomic) SKProduct * product;
-@property (strong, nonatomic) NSString * application;
-@property (strong, nonatomic) NSString * identifier;
-@property (strong, nonatomic) NSArray * stores;
-@property (strong, nonatomic) NSArray * downloads;
-@property (strong, nonatomic) NSString * date;
-@property (strong, nonatomic) NSString * priceLocale;
-@property (strong, nonatomic) NSString * price;
-@property (strong, nonatomic) NSString * currency;
+@property (strong, nonatomic, nonnull) NSString * productName;
+@property (strong, nonatomic, nonnull) NSString * productDescription;
+@property (strong, nonatomic, nonnull) NSString * type;
+@property (strong, nonatomic, nonnull) SKProduct * product;
+@property (strong, nonatomic, nonnull) NSString * application;
+@property (strong, nonatomic, nonnull) NSString * identifier;
+@property (strong, nonatomic, nonnull) NSArray * stores;
+@property (strong, nonatomic, nonnull) NSArray * downloads;
+@property (strong, nonatomic, nonnull) NSString * date;
+@property (strong, nonatomic, nonnull) NSString * priceLocale;
+@property (strong, nonatomic, nonnull) NSString * price;
+@property (strong, nonatomic, nonnull) NSString * currency;
 @property (assign, nonatomic) BOOL active;
 @property (assign, nonatomic) BOOL purchased;