+++ Notificare/NotificarePushLib/Models/NotificareSegment.h
@@ -10,8 +10,8 @@
 
 @interface NotificareSegment : NSObject
 
-@property (strong, nonatomic) NSString * segmentLabel;
-@property (strong, nonatomic) NSString * segmentId;
+@property (strong, nonatomic, nonnull) NSString * segmentLabel;
+@property (strong, nonatomic, nonnull) NSString * segmentId;
 @property (assign, nonatomic) BOOL selected;
 
 @end