+++ Notificare/NotificarePushLib/NotificareEngine.h
@@ -103,4 +103,10 @@ typedef void (^FileUploadResponseBlock)(NSDictionary*file);
 -(NotificareMKNetworkOperation*)fetchAssets:(NSString *)group completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock)errorBlock;
 -(NotificareMKNetworkOperation*)fetchPass:(NSString *)serial completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock)errorBlock;
 
+-(NotificareMKNetworkOperation*)fetchDoNotDisturb:(NSString *)deviceID completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock)error;
+
+-(NotificareMKNetworkOperation*)updateDoNotDisturb:(NSString *)start endDate:(NSString *)end forDevice:(NSString *)device completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock)errorBlock;
+
+-(NotificareMKNetworkOperation*)clearDoNotDisturb:(NSString *)deviceID completionHandler:(NotificationResponseBlock)info errorHandler:(NotificareMKNKErrorBlock) error;
+
 @end